pocproject/requirements.txt
azureml==0.2.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==2.1.15
django-pyodbc-azure==2.1.0.0
djangorestframework==3.10.3
et-xmlfile==1.0.1
feedparser==5.2.1
idna==2.8
jdcal==1.4.1
mysql-connector-python==8.0.18
numpy==1.18.0
openpyxl==3.0.2
pandas==0.25.3
Pillow==6.2.1
psycopg2==2.8.4
pyodbc==4.0.27
python-dateutil==2.8.1
pytz==2019.2
requests==2.22.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
virtualenv==16.7.3
